Add files via upload
This project focuses on developing an AI-powered healthcare website designed 
to revolutionize the way patients access and receive healthcare, particularly for 
those in rural or underserved areas. The primary goal of our website is to 
provide a convenient and accessible platform for individuals to manage their 
health and seek medical advice. In today’s fast-paced world, access to quality 
healthcare is a fundamental human right. However, for many, especially those 
living in rural or underserved areas, this remains a distant dream. The 
challenges of geographical isolation, limited medical infrastructure, and 
financial constraints often hinder timely access to essential medical care